changed getCookie to allow for return dicts and changed ReadingBookStore to use cookies instead of zustand
