+++ src/components/ReadingBooks.jsx
@@ -9,16 +9,15 @@ function ReadingBooks() {
   useEffect(() =>{
     console.log(cookies.user)
     const books = fetchReadingBooks()
-    console.log("after fetch")
-    console.log("ReadingBooks", readingBooks)
   },[])
 
   return (
-    <div>
-      {console.log(readingBooks)}
+    <div className='flex flex-col gap-5'>
       { readingBooks &&
       readingBooks.map((book)=> {
-        return <div>{book.book.title}</div>
+        return <div className='bg-gray-500 min-w-60 min-h-60 border'>
+          <div className=''>{book.book.title}</div>
+          </div>
       })}
     </div>
   )
